Add Validate class, move user input into a separate class.
